update: estilos y agregando funcionalidad al enter del teclado
