+++ README.md
@@ -1,0 +1,18 @@
+# TRABAJO PRÁCTICO – UNIDAD 2
+Tema: IDEs – Control de Versiones – Depuración – CI/CD
+
+__Materia:__ Metodología de Sistemas
+
+__Docente:__  Silvia Viviana Videla
+## Nombre y Apellido del Alumno/a: 
+__Chrestia Luis__ 
+
+__Sandoval Agustín__ 
+
+__Soxkij Nicolás__ 
+
+__Villalba Maximiliano__ 
+
+Curso: Comision 2
+
+Fecha: 27/04/24
